accept array
Summary:
`module.exports = [...]` is a common enough pattern that we need to at least walk the RHS; not doing
so would be grossly inconsistent with walking `module.exports = { ... }`.

The rationale for not doing it so far was to push back on inferring element types that are
unions. That remains a concern, but will be handled by a separate linter that tries to "merge" the
types of elements into a "simple" type (TBD).

Reviewed By: samwgoldman

Differential Revision: D9415266

fbshipit-source-id: c13eed153883d39243b22e324ccbf3e28fb4b85f