accept qualified references on types
Summary:
`M.T` can be used as a type when we have `import type M from ...`. This is what we allow now,
because we want some mechanism to import a module solely for typing but without any runtime
effect. I'm conflicted on whether this is a good idea, especially since there is a workaround
(`import type { T } from ...`). In any case further verification is necessary, but I'm leaving the
enforcement of that case (along with other cases like, e.g., the one that ensures that a variable or
an import that is used as a type is actually a class) for future diffs.

Reviewed By: samwgoldman

Differential Revision: D9437795

fbshipit-source-id: 05b465538b248372b501d9f0e36e78d91bfbdfbf