Change SuperT shape to avoid unnecessary work
Summary:
Currently the SuperT use type contains an objtype and an insttype, but only
cares about the property maps. Creating the larger types requires that we also
create types representing the prototype chain and others.

My larger motivation here is to exploit the fact that extends/mixins can be
fully specialized once inside generate_tests (i.e., during toplevels and super
and implements checks), but needs to be represented as a PolyT over BoundTs
otherwise. Changing check_super to only rely on elements instead of statictype
and insttype makes it easier to exploit this for perf.

Reviewed By: panagosg7

Differential Revision: D9311951

fbshipit-source-id: 680911c7c41cca2dac7ac7c04eceebe30ca7a267