[Ty] Adds ppx_visitors
Summary:
This introduces automatically generated visitors for the `Ty.t` language, using
the [F. Pottier's visitors library](https://gitlab.inria.fr/fpottier/visitors). The benefits of this change:
* custom visitors for each specific operation, rather than a glorified mapper
that just dumps the parts of the output that are not needed in each case -
So now we get: `map`, `iter`, `endo` (mapper that returns the exact input if the
output is invariant), `reduce`, etc.
* automatic generation: no need to maintain the visitor
* bug resilience: I started looking into this as an alternative to debugging a ty_visitor
issue.

The signature file that includes the visitors is a bit verbose, so I'm pasting it here: P59940991,
but not including in the code. The above was provided by Merlin.

In the absence of a `VisitorsRuntime` we need to provide the "base" cases, in
`ty_ancestors.ml` that handle the base cases of the recursion.

The oss build fails when passing both `-package visitors.ppx` and
`-package ppx_gen_rec` on inputs that exercise at least the latter (e.g. `ast.ml`).
Specifically it hits https://github.com/flowtype/ocaml-ppx_gen_rec/blob/bbab5ff0f52910422b2aa6b89d9b58ade5dd700f/src/ppx_gen_rec.ml#L48
For this reason, I'm specifically using
```
<src/common/ty/*.ml*>: package(visitors.ppx)
```
in `_tags`.

Reviewed By: gabelevi

Differential Revision: D9310968

fbshipit-source-id: 90788f4436fda3286c4fee1f8581479e8202d00d