[Fix] Exclude outgoing refinements of conditional expression
Summary:
The outgoing environment of a conditional expression (`e?e1:e2`) needs to
exclude the refinements established when processing the expression `e`. This
already happens in the `if(e){...} else {...}` case.

For example, before
```
var a = { p: 0 };
(a.q ? a.q : "foo");
(a.q: mixed);
```
would only error at the second line. Now it will also error in the third line as well.

This matches the behavior of
```
var b = { p: 0 };
if (b.q) {} else {};
(b.q: mixed);
```
which errors in both lines.

Reviewed By: samwgoldman

Differential Revision: D7164990

fbshipit-source-id: 2b55d497b216e7dcbf7de83ed1c5769d3e10129f