Simplify instance and opaque type representations of type arguments
Summary:
Opaque types and instance types are nominal, which means we can avoid
potentially expensive structural checks. If we compare two such types that have
the same nominal ID, we only need to compare their type arguments, if any.

Because of this, we stored the type arguments along with the instance and
opaque type representation. However, this representation was unnecessarily
complicated.

Furthermore, we stored the tparams_map, which actually includes all type
parameters in scope rather than just those pertaining to the specific type.
Space consumed for this storage was wasted and checking involving those them
was unnecessary.

Finally, the new representation, a list, preserves the order of type parameter
declaration. This is not useful now, but will be useful in a later diff.

Reviewed By: jbrown215

Differential Revision: D9520712

fbshipit-source-id: ab53d3fdd08ba0170a944e0f11934c208c231179