[hack] Option to skip mocking and always use real filesystem in tempfile.ml
Summary:
To reliably test saved states, we want to create and load them in separate
processes (because there is too much hidden global state to safely and reliably
reset them between the operations).

The mocking system doesnt work in this case, because in-process memory it uses
is not shared between processes. Instead, I want to create real temporary directory that
will be used to pass the saved state from one process to another.

Reviewed By: pittsw

Differential Revision: D9322100

fbshipit-source-id: e4ece006cfc33e84690b4f96281fa3a546961c81