Use lint severity at primary location only.
Summary:
The primary location is the first location mentioned in the lint.

The old algorithm worked like this:

1. Find the "severity" at each location mentioned in the lint
2. Use the minimum severity of any location, where `Off < Warn < Err`
3. A `$FlowFixMe` at any location will make the whole error turn off.

{D9323315} then tried modify the logic so that only the "primary location" could turn off a lint (unless there was an explicit comment). Unfortunately, it also made it so that `all=warn` would accidentally tun every lint into an error (the `state == Off` check ignored when `state == Warn`). This was pretty simple to fix.

However, the logic makes even less sense now than before. So I'm simplifying it. Now here's the logic

1. Use the severity at the primary location
2. A `$FlowFixMe` at any location will make the whole error turn off.

This is nice since I think it's what most people want. It's also easier to turn on a lint locally, since before you had to turn on every location mentioned in the lint and now you just have to turn on the primary location.

Reviewed By: wcheng86, samwgoldman

Differential Revision: D9567380

fbshipit-source-id: ece569dafbf3b7c9ccb53c6bdc21457efd4b793a