Avoid flattening severity cover map
Summary:
The lint severity cover data structure is used to find out which lints are enabled, and at which severity level, for a given location. Before this diff, a single `ExactCover.t` could store this information for zero or more files. Many functions which made use of lint severity info expected a single `ExactCover.t` containing severity info for all of the locations which they might query.

However, in the server env, the severity info was stored as a map from file to `ExactCover.t` for that file (and possibly other files in certain cases -- according to @[5401580:Sam Goldman], the `ExactCover.t` for the leader of a cycle would contain severity info for all the files in the cycle). In order to convert this map of `ExactCover.t`s into a single `ExactCover.t` that other functions expected, they would all get unioned together.

This caused serious performance problems because this union operation was `O(N)` on the number of currently-checked files. For non-lazy mode, this is all the files in the project. For lazy mode, it's however many files have ended up getting pulled into the checked set.

To fix this, I have made the un-flattened map get passed around wherever it is needed to avoid the expensive flattening operation. To make this work, I also had to make sure that each `ExactCover.t` represented exactly one file, so that the lint severity info for a given file could be looked up using the file as a key.

Reviewed By: samwgoldman

Differential Revision: D9564121

fbshipit-source-id: a3c8f59b7253af98f819cb8c4b80154ba9d9a821