Add information to RTypeParam reasons for better error messages
Summary:
We use `RTypeParam` for type _argument_ annotations. In addition to the description and location of the `reason_op` used to produce it, we now also propagate the description and location of the reason for the type application in the first place. This enables us to point to why this type argument exists in the first place for missing annotation errors. See the screenshot:

{F136957476}

Reviewed By: gabelevi

Differential Revision: D9500766

fbshipit-source-id: e43917e36603d3831b1583bce2ab50860d3bcd6c