Add traces to EMissingAnnotation errors
Summary:
Missing annotation errors often point to strange places that give no hints
as to how to fix them. Even worse, sometimes these places aren't even in the file that
caused the error to appear.

This diff adds a trace of how we found the missing annotation from the exports, which can be shown using the usual --traces flag.

{F136784334}

As we assert_ground, we collect a stack of reasons. If we find a more specific reason than the
one currently on top of the stack, we replace it. We define "more specific" reasons to be reasons
whose locations are completely contained inside the other reason.

We also use the `def_loc` of reasons for all of the messages in the trace, since those make
where these types are coming from much more clear.

For reference, here is an example of an error that motivated this change: https://flow.org/try/#0CYUwxgNghgTiAEAzArgOzAFwJYHtXwEEAeAFQGcAueMjGLVAcwBp4SA3APgAoAoeeMFAgQARlDABrKl0qsyASngBeDqzZMe86QGUq5RSrUBuHj1CRYCFOmx54AISJ9WAaRABPKjTqMN-EgBqQsggAJKofqxBECEA8sgYkSSxIgBWVADeANQA2hIeem7uALp60SHhAL5JAMJ4GCAAHok83M44aeAYeimpkYLCYpLSbMEgZWPhLPmerh4sHemsvQaqgWPxLfxg9U0YAPx6dagNzRpa8Bl5BXMlEzEgm5Um-Pxm4NBw8KMw8DWZzzebwAtjhgMgICAAHRNAAOOBgGDIygcXBqLAI8iMQA

(see the first error)

The first error mentions Ts, but Ts is nowhere near the place it's pointing at. If it's in a different file that makes this error message even worse.

Reviewed By: gabelevi

Differential Revision: D9438638

fbshipit-source-id: a57f09e87c21f60a272bc14cb062a183b1553797