[PR] Add how to extend the constructor and prototype
Summary:
Addressing issue #4717.
These changes make it clear how to extend the prototype and constructor of a class. Methods added in the class declaration do not need to be annotated. At this time, methods and properties defined on the constructor (static) and on the prototype need to be annotated in the class definition. This is probably a best practice since, if the class is modified in a location not local to the definition we will want the type checker to notify us.
Pull Request resolved: https://github.com/facebook/flow/pull/4719

Reviewed By: gabelevi

Differential Revision: D9585833

Pulled By: mrkev

fbshipit-source-id: 421cb2ca4ed489848fe0becf518a3ca96a2ef4d3