[type-visitor] Shallowly inspect PolyTs for EvalTs
Summary:
This change shallowly inspects the PolyTs for EvalTs and does what we ought to do in that case. This is still overly conservative, but does get rid of some of the remaining errors.

We recently made a change to `createFragmentContainer` that changed a `$` to `.`. We did this because the type arguments to builtins should be visited from a constant positive position, but the type alias added some misdirection which caused the constant positive position to be forgotten. This fix would have also fixed many of those missing annotation errors.

Reviewed By: samwgoldman

Differential Revision: D9296921

fbshipit-source-id: a2e036f013e6378ece187e50e90bba2da905cf23