Get your sqlite3 chocolate out of my NO_SQLITE3 peanut butter
The Hack project uses some sqlite3 integration which Flow does not use.
We introduced NO_SQLITE3 to conditionally compile that stuff for Hack,
but the define is granular and Hack changes break Flow builds pretty
regularly. This is especially complicated because Flow builds on Windows
and Hack doesn't.

Now that sqlite3 code is (mostly) untangled from the sharedmem code,
it's easier to ignore the entire thing with a big #ifndef block around
the sqlite3 code.