[hack] Change hh_header_t to bitmask with tagged low-order bit
Summary:
In a follow-up, I replace the copying collector with a compacting collector.
The compacting collector needs to walk the heap and distinguish between headers
and aligned pointers. Since aligned pointers will always have the low-order bit
unset, this is sufficient.

Only separating this change out to reduce noise in the follow-up.

Reviewed By: gabelevi, pittsw

Differential Revision: D9628788

fbshipit-source-id: d5469bc6fec407ab80a9a27dc36c4c12f40fe81b