[Refactor] Move subst into Type_mapper.ml
Summary:
This diff moves subst into Type_mapper to reduce the LOC in flow_js.ml. Up next, I'll be changing the signature of the type_mapper
to factor out all methods that interact with context into separate functions, since those are the ones that require the most thought. I'll also
be making the class abstract.

These changes are part of a larger change to use the Type_mapper (instead of the visitor) in the context optimizer, which will allow
us to perform type-level optimizations on our types in sig_cx. Additionally, it'll allow us to mark tvars created there derivable, which
will get rid of errors in  D7169498.

Reviewed By: samwgoldman

Differential Revision: D9684571

fbshipit-source-id: 04ea3e4f253829b379d7d0cf1da26323505c2250