[Refactor] Make type mapper pure
Summary:
This factors out the context-modifying parts of mapping into abstract methods. This forces users to think hard about how the context should be modified, if at all.
Since subst used all of the context modifying parts of the type mapper, we've moved their implementations there.

Up next: modify the signature of Type_mapper to make sure the methods exposed are the same as those in Type_visitor

Reviewed By: samwgoldman

Differential Revision: D9686862

fbshipit-source-id: 08e7b21dcc45e5aa32e501b229d6866565fe29a3