Remove ability to use functions as types
Summary:
This diff kills some old code that was designed to support code that predates
classes. This code reflects the old Flow pragmatism where best-effort type
checking was better than nothing. As a result, the logic to support this has a
lot of holes which reduce coverage.

Now that classes have been around for a long time, we should drop support for
this pattern. This will make it an error to use a function in a type position.

Reviewed By: panagosg7

Differential Revision: D9725084

fbshipit-source-id: 3b5b0a9c22abb1a53e85d7d31e29f58284fd18fe