Profile worker wall time RUN/READ/SEND/IDLE
Summary:
We don't have much insight into worker utilization or the overhead of having large worker requests/responses. This tries to remedy that. Now, whenever someone calls `MultiWorkerLwt.call`, we track

* Run wall time. The time spent running a job on a worker (excluding reading a request and sending a response).
* Read wall time. The time spent reading a request from the controller
* Send wall time. The time spent sending a response to the controller
* Idle wall time. The time spent waiting for the next job or waiting for the last few jobs to finish.

Only time inside of `MultiWorkerLwt.call` is tracked. If you add a `sleep` outside of the `MultiWorkerLwt.call`, that won't affect these new stats.

The absolute wall times are not terribly easy to understand, since we have N workers and all the metrics are the sum of all their wall times. So the easiest way to understand these metrics are as percentages.

Reviewed By: samwgoldman

Differential Revision: D9779816

fbshipit-source-id: 18b5f36fe88bad2ee55037d97ff56ab950a28dba