Tweak GC settings
Summary:
This seems to save 10-40s off of a total check. Most of those
wins are from `Merge`, though any step which GC's a lot seems to be
faster. The main downside to GCing less often would be using more
memory, but cgroups don't seem to detect any issue.

{D9826945} now splits out major and minor GC'ing in the table, which helps with intuition. Before this change, Merge workers would spend ~5% of the time doing minor collections and ~13% of the time doing major collections. After this change, it drops to 2% and 3%.

We expect minor collections to be less frequent, since the minor heap is larger. We expect major collections to be less frequent and cheaper since fewer objects will be falsey promoted to the major heap (since they can stay in the minor heap longer).

Reviewed By: samwgoldman

Differential Revision: D9827076

fbshipit-source-id: e5fcfdd14982dae149b4b87a31e0a38adc47f356