Drop trailing comma in array/object pattern because trailing rest
Summary:
In the layout printer do not add a trailing comma to array/object patterns. This prevents syntax errors with rest elements in array patterns and rest properties in object patterns.

Note that while object pattern rest _can_ have a trailing comma, for a long time leading up to the final spec it wasn't allowed and many tools still need to catch up to match that syntax. As such it's best for us to follow suit and not print the trailing comma.

Two test cases were added that assert no comma is added to the end of a multi-line obj/arr pattern that ends with a rest.

Reviewed By: mroch

Differential Revision: D9850675

fbshipit-source-id: b48fcd789b8843ce982c835663a7984673914aa1