Update Replacement_printer to use the diff format stored in the DiffHeaps
Summary:
This diff promotes Replacement_printer to public (yay!) to be in a new target under the printers folder.

It uses the same format as the DiffHeaps introduced by the previous diff on the stack, and is split in 2 steps:

mk_patch is safe to run in workers as they don't modify the file.

Reviewed By: dsainati1

Differential Revision: D9874006

fbshipit-source-id: ff2000f89927c6b68579872e55edaff53503ef9c