Use LSP error structure not exn
Summary:
This diff moves towards using Lsp.Error.t structure for errors, rather than ocaml exceptions.

(Why? because there will be places where I want to marshal an error, and in ocaml if you unmarshal an exception then you can no longer match on it).

I haven't moved 100% towards using error.t because that'd be a bunch of rewriting in hack-lsp that I can't be bothered to do right now...

Reviewed By: arxanas

Differential Revision: D9829917

fbshipit-source-id: 431045b30f54eedc99014a58a101646acb9c109b