flow cap errors at 200 for Nuclide
Summary:
If an LSP server reports too many errors, then it (1) takes a long time to send them over the wire, (2) takes a long time for Nuclide to unmarshal them and do its internal algorithms and then display them. The net effect is that Nuclide is completely frozen for a few seconds at a time. Visual Studio is the same -- it gets pretty sluggish with lots of errors, so they capped at 100. (Visual Studio Code is apparently pretty zippy in this respect however).

This diff caps the number of errors at 200 per file. If there are any more, then we'll introduce an additional diagnostic at the top line of the file saying "[Only showing 200/895 diagnostics]". My thinking is that if there are a load more errors in the file, then they're generally cascading errors where error recovery didn't work, or they all speak to a common root cause and don't add value to users. I picked the number 200 from half way between the observation that 100 remains perfectly responsive in Nuclide, and 1000 feels noticeably unresponsive.

If we do cap, I wanted to sort the list so only the first 200 errors are shown. This sort might take a bit of time, but I don't think there are good alternatives. (Still, doing the sort on the devserver is quicker than sending it all to the client and having the sort done there!)

Reviewed By: pakoito

Differential Revision: D9951918

fbshipit-source-id: fcaddea386d79f399d122741abe31aed5f4db14b