flowLSP don't send go-to-def results when there weren't any
Summary:
When GetDef_js.get_def is unable to find the definition (e.g. when it was asked to find definition for a symbol in a comment) then it returns Loc.none

This should be returned as an empty array to the LSP textDocument/definition request. That's what this diff does.

(Previously, in Nuclide, when you cmd+hovered then it would underline every symbol because every symbol gave a definition response, and when you cmd+clicked then it would take you to the top of the file which corresponds to line=-1 which is what Loc.None is).

Reviewed By: nmote

Differential Revision: D9977619

fbshipit-source-id: 4e1a3bc401c183da000970b7b27be4b276c1fa89