Switch saved states from `Flow_version` to `Flow_build_id`
Summary:
We only bump the `Flow_version` on releases. This means during
development, we may have a binary with the same `Flow_version` but a
different binary. This can lead to bad things when we try to unmarshal
data from a saved state file.

To guard against this, we should use `Flow_build_id`, which asserts that
the binary is the exact same binary.

Reviewed By: jbrown215

Differential Revision: D9663612

fbshipit-source-id: 37ff6ef07a6555e002e90eb1117f70bc0dda034b