Ignore "propTypes" objects
Summary:
This diff introduces an exception case in the signature verifier for `static propTypes` members of a class.

Currently, `propTypes` cause an error in signature validation as they're considered part of the external interface. After an internal discussion, we agreed that a mechanism used to narrow down some internal state shouldn't be considered exported by default.

We're aware that there are several instances where `propTypes` are consumed outside of the module exporting them, and we'll have to fix those manually :)

Reviewed By: mroch

Differential Revision: D9940441

fbshipit-source-id: 385c41ac77e9c679238b368f0d0212177be4f12d