Add ALoc (Abstract Location) module
Summary:
We would like to prevent rechecks from occurring when only locations are changed. However, locations are used in type errors downstream. If we simply omitted them when computing the sighash, we would end up reporting errors with stale locations.

Instead, we would like to use abstract locations to describe the locations of things which can be exported. They will have stable keys and we can then do a lookup to get the concrete location when it comes time to render user output.

Right now, this module just hides the implementation of `Loc.t`. Later on, though, we will be able to use it to manage abstract locations as described above.

Reviewed By: samwgoldman

Differential Revision: D9980406

fbshipit-source-id: d5bc94f3a62689a32a67b43eadad33d30e9f6e32