Convert Reason locs to ALoc.t
Summary:
See D9980406 for an overview.

This diff changes the type of the `loc` field in `Reason.t` to `ALoc.t`, and then makes the required changes to get Flow to build.

The current locations of `ALoc.of_loc` and `ALoc.to_loc` are not what we want them to be long term. I'll incrementally push the boundaries of these function calls so that `ALoc.t` is used everywhere it needs to be.

Reviewed By: samwgoldman

Differential Revision: D9980408

fbshipit-source-id: b37112afd3605f55025c8855f8362dc4b167bfda