Switch saved states from `Flow_version` to `Flow_build_id`
Summary:
We only bump the `Flow_version` on releases. This means during
development, we may have a binary with the same `Flow_version` but a
different binary. This can lead to bad things when we try to unmarshal
data from a saved state file.

To guard against this, we should use `Flow_build_id`, which asserts that
the binary is the exact same binary.

Reviewed By: jbrown215

Differential Revision: D9663612

fbshipit-source-id: 7e03b600b9f1c77ba920824d8f47336462542d94