[Refactor] Use TypeMapper for ContextOptimizer
Summary:
Instead of using the Type_visitor, now we use the Type_mapper for the context optimizer. This allows us to actually change the types in the sig_cx.

This diff should maintain the same behavior as what there was before. In a follow up diff, I'll be making stable OpenTs derivable.

Important note: We're not making any new ids here for anything we store in the context. Instead, we are updating the stored entries in the context. This is to make sure these changes are propagated everywhere a single id may have been visited and to maintain that equal ids are still equal after mapping.

Reviewed By: samwgoldman

Differential Revision: D9722062

fbshipit-source-id: e3a3c63ab81d7c6ecd3b1b258936fe6ae455a0e6