flow respects cancellation
Summary:
This diff provides "best-effort" cancellation support for FlowLSP.

The way LSP cancellation works is that Nuclide asks for a request to be cancelled. The server is at liberty to respect or ignore it.

In this diff, lspCommand will receive the cancellation message. It will pass it via a pipe into the monitor. The monitor will get round to handling it and stick it in a pipe to the server. The server will get round picking up the message from its stdin and will add it immediately to a "requested cancellations" set (vs normal requests which get added to a queue of pending workitems).

When the server comes round to handling a workitem, it first checks whether cancellation for this workitem has previously been requested. If so it will decline to handle the workitem.

"Best Effort": if the server has already embarked upon processing the workitem then it will ignore it; if the cancellation request happened to be going through any of the pipes at the time processing starts then it will be ignored; if the server ended up doing a load of work on its primary thread (rather than shunting off work to its workers) then it will be so busy it likely never gets to drain its incoming pipe and so won't see the cancellation request in time.

Improving any of these things would take a lot of effort that doesn't seem worth it.

Reviewed By: gabelevi

Differential Revision: D9830248

fbshipit-source-id: 6bcc24bd19189e355b32ef46f957f6284af56256