disable test that was causing AppVeyor flakiness
Summary:
I'm disabling this test because it's flakey on AppVeyor.

What happens 1 in 50 runs is that 'flow force-recheck --no-auto-start .flowconfig' sends a message to the monitor and thence the server telling it to force-recheck, and it's a race whether monitor would be able to send the response back to forceRecheckCommand before dying or not.

If it happens to die before it can send a response then forceRecheckCommand will use its retry logic to send the FORCE_RECHECK request a second time.

And if lspCommand happens to restart the server before that retry request is sent, then the retry request will end up killing the newly started server!

I'm disabling the test for now so it doesn't interfere with AppVeyor.

In any case, it won't have a bad user experience - the retry behavior of forceRecheckCommand doesn't correspond to any real watchman behavior; and if the user does happen to do forceRecheckCommand in a way that stops flow, then the worst that will happen is that Nuclide pops up a box saying "flow is stopped [restart]".

Reviewed By: gabelevi

Differential Revision: D9985572

fbshipit-source-id: 37545d6b0a9c3c22c0160742f02f63957e62a43f