Tidy up: it's monitor's job to send Prot.ServeExit
Summary:
lspCommand needs to know when the server died - was that death from one of a small number of known whitelisted reasons that merit an automatic restart? or was it for some other reason that doesn't?

D5798983 introduced Prot.ServerExit, sent from server to client (via monitor) from inside rechecker.ml for the cases where the server is about to die.

But that had race conditions - what if the server sends that message to monitor, and dies, and monitor sees that the server has died and so dies itself BEFORE it forwards the message on to the client?

D7616745 fixed the race by making the monitor be the party responsible for catching the server's exit code and sending that on to the client, and then closing the connection to the client. It was therefore able to remove the code in sender which sent that Prot.ServerExit message.

Except it didn't quite clean up all the way. This diff removes the last remaining place where server attempts to send the Prot.ServerExit message.

Reviewed By: gabelevi

Differential Revision: D9985425

fbshipit-source-id: 678fc1e55d989780ba6ab511f51cbdfbefe4f212