[hack] Add a command line flag to overwrite hh.conf config values.
Summary:
A lot of times when testing things I'll be trying to compare different values for config settings in hh.conf, which means manually updating the file or copying around various versions of it or running with a diff on top of my local stack that hardcodes a value.  It's much easier to just give these values as command line arguments.

Since this feature is primarily for debugging, I don't have any checking in the server monitor that the arguments given match those given to an existing server if no server needs to be started.

Differential Revision: D10019212

fbshipit-source-id: 2b86813f661c848b79f7751e8806eb4d32cdb7bf