Fix test mapper
Summary: This makes it so that type annotation nodes have physical equality preserved if they aren't changed, an invariant that our differ relies upon to give reasonable results. This has not effect on the tests as they are, but it tripped me up when I was testing one of my next diffs.

Reviewed By: dsainati1

Differential Revision: D10111388

fbshipit-source-id: 27e0326ddbe75cf7c8a68a993f3ca5fbf0916373