special case MaybeT and OptionalT when they appear as upper bounds on unions
Summary: We have a special case to deal with a flow between two union types, but when flowing a union type to a maybeT or an optionalT we split the union open regardless, causing exponential blowup when comparing optional unions. This adds cases to treat maybeT<T> as (void | null | T) and optionalT as (void | T) when these types appear as upper bounds on unions.

Reviewed By: samwgoldman, avikchaudhuri

Differential Revision: D10088520

fbshipit-source-id: 576506573e7c6e23e3299e25d0b3b998bc661256