remove Break_always mode in favor of Newline
Summary: `Newline` inserts a newline even in compact mode. previously the only way to do this was with a `Break_always` sequence with either the left or right inline mode enabled depending on whether you wanted a prefix or suffix newline. now you can just put a `Newline` where you want it, and we don't need `Break_always`.

Reviewed By: pieterv

Differential Revision: D10092542

fbshipit-source-id: 3a9360fbd00125156032718d7229491ad0ba074f