add Group and Indent layout nodes
Summary:
adds a node that increases the indentation of its children by 2 spaces if the code doesn't fit on one line.
also adds a `Group` node that is like `Sequence ~break:Break_if_necessary`, but doesn't indent or add newlines. combined with `Newline` and `Indent`, we can replace `Sequence` in a more configurable way.

Reviewed By: pieterv

Differential Revision: D10112531

fbshipit-source-id: 57cb432849253078512f3c3333f173731516ae97