[Typed AST] Adds support for JSX
Summary:
Adds support for nodes in the `Ast.JSX` module.

Similar to the general design of the typed AST, construction follows the typechecking
traversal, so for nodes that are never reached we return "error" nodes (e.g. see `jsx_title`).
To aid practicality, a small exception to this is the case of the closing tag of a JSX node:
```
<m> .. </m>
         ^
```
Here, leveraging the invariant that the closing tag will match the opening, we map the
types inferred for the opening tag, keeping the locations of the closing tag. This happens
in `jsx_match_closing_element`.

JSX titles are checked after being translated to Member expressions (through
`jsx_title_member_to_expression`). We recover the initial structure by reversing this
process in `expression_to_jsx_title_member`.

Addresses: https://github.com/facebook/flow/issues/6900

Reviewed By: pakoito

Differential Revision: D10117360

fbshipit-source-id: 9bc76cf4651092c469c4f3a304c3dad79f0d8542