Add `ALoc.source`
Summary:
Whatever abstract representation we come up with for locations, it will definitely include the source. That's good, because there are all sorts of places where it's used and we don't want to have to do a lookup in the map in each of these places.

For now, it's just backed by `Loc.source`.

Reviewed By: samwgoldman

Differential Revision: D10062955

fbshipit-source-id: 6aec56383baeaa2e9a5d2dad7b03ea9a56ea7a33