FlowLSP bug reports, if they can't find logfiles, will at least report filenames
Summary:
The current way that flowLSP's telemetry/rage command works is it tries to include the server log and the monitor log, and the old versions of both. But if it can't find those files, then it simply doesn't mention them in the bug report.

This diff changes it so that if the file doesn't exist, then the bug report at least mentions the name of the file that it tried to read.

When testing this I observed that I had a bug when trying to fetch the .old logfiles. I'd used the wrong Path.concat function, so it was looking for "xyz.log/.old" rather than "xyz.log.old". This diff corrects that problem.

Reviewed By: nmote

Differential Revision: D10189400

fbshipit-source-id: 3a5a9ffe39f8773dd153f7b108f439a19e2c2f6b