Remove AssertRestParamT
Summary:
This use type is designed to validate a definition where the annotation of a
rest parameter is invalid. However, this constraint causes crashes if it is
created before substitution, which currently happens in declared classes.

Since any call of such an invalid function declaration will naturally expose
this error, we take the simplest path, and just eliminate the use type
entirely.

Reviewed By: avikchaudhuri

Differential Revision: D10204220

fbshipit-source-id: bbf07f2bec6bbfc1cc91a29ff5f902657ab25c25