Rewrite/fix class layout
Summary:
I think there's something fishy going on with `list`, which generates two levels of `Sequence`, when the wrapped code is exactly 80 cols (or maybe 79?), and it gets indented. i'm not exactly sure what happens but the result is that it prints on one line, but with a space missing, which I think happens because it thinks it's wrapping but doesn't.

either way, `Sequence` is confusing and that's why I'm trying to remove it. so this diff replaces its caller in class layout with `Group`, which fixes the test provided by panagosg7

Reviewed By: pieterv

Differential Revision: D10205357

fbshipit-source-id: b823f84340816e3e67b594cabf252196fc38525f