signature verification error for empty object literal
Summary:
An empty object literal is unsealed in Flow, which means that we allow adding properties to
it. Demanding an annotation is better than pretending that it has the empty object type.

Note that the type Flow would infer would have non-optional properties, but such an annotation will
not currently work: the only way to make this work currently is by using optional properties. In
follow-up work we should make the inferred annotation work for consistency, and deal with the
unsoundness of this type (which already exists in our inference) separately.

Reviewed By: panagosg7

Differential Revision: D10158508

fbshipit-source-id: b6f6517ef24a461298e6d9b42e986b6e1fb3a74a