audit all wildcard record labels in generator
Summary:
Ignoring record labels in ASTs is a risk for signature generation, since we might not preserve key
information when mapping definitions to their corresponding declarations. Some related bugs were
fixed earlier in this stack, so this diff audits all possible sources of such bugs.

The only TODO unraveled by this diff is to investigate whether we are properly handling async /
generator functions.

Reviewed By: samwgoldman

Differential Revision: D10187323

fbshipit-source-id: b0fc6f4ab4a753c8ac91a1be800abc2fb12dece5