Introduce new helper function
Summary:
Several callsites to `diff_if_changed_opt` wrapped their arguments in `Some` simply because the right variant of `diff_if_changed` did not yet exist.

This diff adds the `diff_if_changed_ret_opt` variant, for use when the arguments are non-optional but the supplied function returns an option.

Differential Revision: D10261505

fbshipit-source-id: 74c23692d465c3655c49cabda187ecb370a99e7a